Add namespacing to package.
* Use PSR-4 autoloader in composer.json.
* Move the parser class into the src directory under the namespace joshfraser.
* Update all usages of the parser to use the namespace.
* Include composer.lock file.
* Define the license in composer.json; Apache-2.0 was the license already specified in the upstream.
