+++ assets/style.css
@@ -18,4 +18,6 @@ tr{ page-break-inside: avoid; }
 .unit-tests tbody .pass:hover { background:#ecffc1; color:#7ec43c; }
 .unit-tests tbody .fail:hover { background:#ffc1c1; color:#e53d3d; }
 .unit-tests tbody .pass td:hover { background:#c5f458; color:#8ec100; }
-.unit-tests tbody .fail td:hover { background:#f25959; color:#c10000; }
+.unit-tests tbody .fail td:hover, .unit-tests tbody .debug-expected td { background:#f25959; color:#c10000; }
+.unit-tests tbody .debug-expected td:first-child { font-weight: bold; }
+.unit-tests tbody .debug-expected td { color: #222222; }
