+++ tests/phpunit.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<phpunit bootstrap="../parser.php"
+<phpunit bootstrap="../vendor/autoload.php"
          backupGlobals="false"
          backupStaticAttributes="false"
          strict="true"
